% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Matrix_to_SparseM}
\alias{Matrix_to_SparseM}
\title{Sparse Matrix Conversion: Matrix to SparseM}
\usage{
Matrix_to_SparseM(X)
}
\arguments{
\item{X}{A sparse matrix from the Matrix package}
}
\value{
A sparse Matrix from the SparseM package ("matrix.csc")
}
\description{
Sparse Matrix Conversion: Matrix to SparseM
}
