% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{prune_dfm}
\alias{prune_dfm}
\title{Feature matrix pruning}
\usage{
prune_dfm(dfm, minimum_threshold = 0.005, maximum_threshold = 1)
}
\arguments{
\item{dfm}{a sparse matrix (Matrix)}

\item{minimum_threshold}{minimum document frequency threshold for relative pruning}

\item{maximum_threshold}{maximum document frequency threshold for relative pruning}
}
\value{
A pruned sparse Matrix
}
\description{
Prunes a document feature matrix (dfm) using relative thresholds
}
\examples{
# example feature matrix
m <- Matrix(round(replicate(10, abs(rnorm(20))) * 10))
colnames(m) <- as.character(1:10)
dim(m)
n <- prune_dfm(m)
dim(n)

}
