% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{rmsd}
\alias{rmsd}
\title{Root mean-square deviation}
\usage{
rmsd(y1, y2)
}
\arguments{
\item{y1}{numeric data series 1}

\item{y2}{numeric data series 2}
}
\value{
numeric RMSD value
}
\description{
Root mean-square deviation
}
