% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slam_to_Matrix}
\alias{slam_to_Matrix}
\title{Sparse Matrix Conversion: slam to Matrix}
\usage{
slam_to_Matrix(X)
}
\arguments{
\item{X}{A sparse matrix from the slam package}
}
\value{
A sparse Matrix from the Matrix package
}
\description{
Sparse Matrix Conversion: slam to Matrix
}
