% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{tmca_irr_statistics}
\alias{tmca_irr_statistics}
\title{Inter-rater agreement statistics}
\usage{
tmca_irr_statistics(pred, labels)
}
\arguments{
\item{pred}{vector (factor) of predicted labels}

\item{labels}{vector (factor) of true labels}
}
\description{
Kappa and alpha for irr
}
\examples{
random1 <- sample(0:1, 1000, replace = T)
random2 <- sample(0:1, 1000, replace = T)
tmca_irr_statistics(random1, random2)
}
