% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{tmca_proportions}
\alias{tmca_proportions}
\title{Proportional classification evaluation}
\usage{
tmca_proportions(predicted_labels, true_labels, facets = factor(),
  positive_class = NULL, verbose = F)
}
\arguments{
\item{predicted_labels}{vector (factor) of predicted labels}

\item{true_labels}{vector (factor) of true labels}

\item{facets}{factor with data facets to group labels}

\item{positive_class}{name / level of the positive class}
}
\value{
list with evaluation metrics: root-mean-square deviation, Pearson's r, and predicted/true proportions per facet
}
\description{
Proportional classification evaluation
}
